% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_unique_edge}
\alias{add_unique_edge}
\title{Add an undirected edge from v1 to v2 to graph g if it does not already exist.}
\usage{
add_unique_edge(g, v1, v2)
}
\arguments{
\item{g}{Graph representing social network}

\item{v1}{First node in edge pair}

\item{v2}{Second node in edge pair}
}
\value{
igraph Graph
}
\description{
Add an undirected edge from v1 to v2 to graph g if it does not already exist.
}
\examples{
# Add one unique edge between nodes 1 and 4 to empty ten-node network
g <- igraph::make_empty_graph(n = 10)
g <- add_unique_edge(g, 1, 4)

}
