% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{AgentBasedModel}
\alias{AgentBasedModel}
\title{AgentBasedModel class}
\description{
Manages agent initialization, graph structure, and social connections.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AgentBasedModel-new}{\code{AgentBasedModel$new()}}
\item \href{#method-AgentBasedModel-get_agent}{\code{AgentBasedModel$get_agent()}}
\item \href{#method-AgentBasedModel-clone}{\code{AgentBasedModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$new(n_agents = NULL, graph = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-get_agent"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-get_agent}{}}}
\subsection{Method \code{get_agent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$get_agent(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
