% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\docType{data}
\name{contagion_learning_strategy}
\alias{contagion_learning_strategy}
\title{Define contagion learning "strategy".}
\format{
An object of class \code{LearningStrategy} (inherits from \code{R6}) of length 7.
}
\usage{
contagion_learning_strategy
}
\description{
Define contagion learning "strategy".
}
\keyword{datasets}
