% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\docType{data}
\name{contagion_model_dynamics}
\alias{contagion_model_dynamics}
\title{Define contagion learning "strategy".}
\format{
An object of class \code{ModelDynamics} (inherits from \code{R6}) of length 7.
}
\usage{
contagion_model_dynamics
}
\description{
Define contagion learning "strategy".
}
\keyword{datasets}
