% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{contagion_partner_selection}
\alias{contagion_partner_selection}
\title{Contagion-based partner selection}
\usage{
contagion_partner_selection(learner, model)
}
\arguments{
\item{learner}{An Agent instance.}

\item{model}{An AgentBasedModel instance.}
}
\value{
A single neighbor Agent.
}
\description{
Selects one neighbor at random for potential contagion interaction.
}
\examples{
model <- make_abm(graph = igraph::make_graph(~ 1-2), adoption_rate = 1.0)
learner <- model$get_agent(1)
teacher <- contagion_partner_selection(learner, model)
}
