% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{DEFAULT_PARAMETERS}
\alias{DEFAULT_PARAMETERS}
\title{Default parameters to create an agent-based model.}
\format{
An object of class \code{ModelParameters} (inherits from \code{R6}) of length 13.
}
\usage{
DEFAULT_PARAMETERS
}
\description{
Default parameters to create an agent-based model.
}
\keyword{datasets}
