% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{dummy_learning_strategy}
\alias{dummy_learning_strategy}
\title{Make a dummy learning strategy for mockups and testing.}
\usage{
dummy_learning_strategy()
}
\description{
Make a dummy learning strategy for mockups and testing.
}
