% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{dummy_model_dynamics}
\alias{dummy_model_dynamics}
\title{Make a dummy learning strategy for mockups and testing.}
\usage{
dummy_model_dynamics()
}
\description{
Make a dummy learning strategy for mockups and testing.
}
