% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{fixated}
\alias{fixated}
\title{Predicate function: has the model fixated on a single behavior?}
\usage{
fixated(model)
}
\arguments{
\item{model}{An AgentBasedModel}
}
\value{
TRUE if all agents have the same behavior
}
\description{
Predicate function: has the model fixated on a single behavior?
}
\examples{
net <- igraph::make_graph(~ 1-2)
model <- make_abm(graph = net)
trial <- run_trial(model, stop = fixated) # <- "stop trial when fixated"
}
