% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{frequency_bias_interact}
\alias{frequency_bias_interact}
\title{Interaction function for frequency-biased adaptive learning.}
\usage{
frequency_bias_interact(learner, ., model)
}
\arguments{
\item{learner}{Agent currently selected as learner.}

\item{.}{Not used (no teacher for frequency bias).}

\item{model}{The ABM instance}
}
\description{
Interaction function for frequency-biased adaptive learning.
}
