% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\docType{data}
\name{frequency_bias_learning_strategy}
\alias{frequency_bias_learning_strategy}
\title{Define frequency-biased learning strategy.}
\format{
An object of class \code{LearningStrategy} (inherits from \code{R6}) of length 7.
}
\usage{
frequency_bias_learning_strategy
}
\description{
Define frequency-biased learning strategy.
}
\keyword{datasets}
