% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\docType{data}
\name{frequency_bias_model_dynamics}
\alias{frequency_bias_model_dynamics}
\title{Define frequency-biased learning strategy.}
\format{
An object of class \code{ModelDynamics} (inherits from \code{R6}) of length 7.
}
\usage{
frequency_bias_model_dynamics
}
\description{
Define frequency-biased learning strategy.
}
\keyword{datasets}
