% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{frequency_bias_select_teacher}
\alias{frequency_bias_select_teacher}
\title{Frequency biased teacher selection does nothing}
\usage{
frequency_bias_select_teacher(agent, model)
}
\description{
Frequency biased teacher selection does nothing
}
