% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{G_NM}
\alias{G_NM}
\title{Erdős-Rényi random graph G(N, M).}
\usage{
G_NM(N, M)
}
\arguments{
\item{N}{number of nodes/agents}

\item{M}{number of edges to be randomly assigned}
}
\value{
igraph Graph instance
}
\description{
Erdős-Rényi random graph G(N, M).
}
\examples{
# Create a 10-node network with 10 randomly-assigned edges
library(igraph)
net <- G_NM(10, 10)
plot(net)

}
