% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_all_possible_edges}
\alias{get_all_possible_edges}
\title{Get all possible edges between node indices 1 to N for either directed or
undirected networks.}
\usage{
get_all_possible_edges(N, directed = FALSE)
}
\arguments{
\item{N}{number of nodes}

\item{directed}{Whether or not the possible edges are for directed graphs}
}
\value{
table of node pairs representing edges
}
\description{
Get all possible edges between node indices 1 to N for either directed or
undirected networks.
}
\examples{
# Get a table of vertex pairs representing possible edges with ten vertices.
Epossible <- get_all_possible_edges(10)

}
