% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ggnetplot}
\alias{ggnetplot}
\title{Command that operates like ggplot}
\usage{
ggnetplot(net, layout = NULL, ...)
}
\arguments{
\item{net}{Network to plot}

\item{layout}{Optional igraph-style plotting layout}

\item{...}{Additional key-value aesthetics to pass to aes() in ggplot call}
}
\description{
Command that operates like ggplot
}
\examples{
library(ggnetwork)
socnet <- igraph::make_graph(~ 1-2:3:4,2:4)
V(socnet)$name <- c("Mateus", "Marcos", "Lucas", "João")
ggnetplot(socnet) + 
  geom_edges(linewidth=0.1) + 
  geom_nodes(color = "#008566", size=3) + 
  geom_nodelabel_repel(aes(label = name), size = 1.5) + 
  theme_blank()
}
