% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{iterate_learning_model}
\alias{iterate_learning_model}
\title{A generic method for iterating a learning model, setting the current
behavior and fitness to be whatever was identified as the next behavior
and fitness.}
\usage{
iterate_learning_model(model)
}
\value{
NULL Operates in-place to update all agent's behavior if necessary.
}
\description{
A generic method for iterating a learning model, setting the current
behavior and fitness to be whatever was identified as the next behavior
and fitness.
}
