% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{learning_model_step}
\alias{learning_model_step}
\title{More modern v0.1 alias for iterate_learning_model.}
\usage{
learning_model_step(model)
}
\description{
More modern v0.1 alias for iterate_learning_model.
}
