% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_strategy.R
\name{LearningStrategy}
\alias{LearningStrategy}
\title{LearningStrategy}
\description{
LearningStrategy

LearningStrategy
}
\details{
Represents a reusable learning strategy composed of a partner selection function,
an interaction function, and a descriptive label.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearningStrategy-new}{\code{LearningStrategy$new()}}
\item \href{#method-LearningStrategy-get_partner_selection}{\code{LearningStrategy$get_partner_selection()}}
\item \href{#method-LearningStrategy-get_interaction}{\code{LearningStrategy$get_interaction()}}
\item \href{#method-LearningStrategy-get_label}{\code{LearningStrategy$get_label()}}
\item \href{#method-LearningStrategy-clone}{\code{LearningStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearningStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearningStrategy-new}{}}}
\subsection{Method \code{new()}}{
Create a new LearningStrategy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearningStrategy$new(partner_selection, interaction, label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{partner_selection}}{Function to select the teacher (or NULL)}

\item{\code{interaction}}{Function to define the interaction (or NULL)}

\item{\code{label}}{A string label for this strategy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearningStrategy-get_partner_selection"></a>}}
\if{latex}{\out{\hypertarget{method-LearningStrategy-get_partner_selection}{}}}
\subsection{Method \code{get_partner_selection()}}{
Get the partner selection function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearningStrategy$get_partner_selection()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearningStrategy-get_interaction"></a>}}
\if{latex}{\out{\hypertarget{method-LearningStrategy-get_interaction}{}}}
\subsection{Method \code{get_interaction()}}{
Get the interaction function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearningStrategy$get_interaction()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearningStrategy-get_label"></a>}}
\if{latex}{\out{\hypertarget{method-LearningStrategy-get_label}{}}}
\subsection{Method \code{get_label()}}{
Get the strategy label
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearningStrategy$get_label()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearningStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearningStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearningStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
