% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{load_igraph_from_csv}
\alias{load_igraph_from_csv}
\title{Load an Undirected igraph Object from a CSV Edge List}
\usage{
load_igraph_from_csv(csv_file)
}
\arguments{
\item{csv_file}{Path to a CSV file with two columns representing edges.}
}
\value{
An undirected igraph object.
}
\description{
Loads a graph from a CSV file containing a two-column edge list.
Assumes columns are either \code{from} and \code{to}, or \code{source} and \code{target}.
}
\examples{
fnet <- load_igraph_from_csv("marketville-friends-coleman-feld.csv")

}
