% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_strategy.R
\name{make_learning_strategy}
\alias{make_learning_strategy}
\title{Factory function for creating a LearningStrategy}
\usage{
make_learning_strategy(partner_selection, interaction, label)
}
\arguments{
\item{partner_selection}{A function to select a teacher (or NULL)}

\item{interaction}{A function defining interaction (or NULL)}

\item{label}{A string label for this strategy}
}
\value{
A \code{LearningStrategy} instance
}
\description{
Factory function for creating a LearningStrategy
}
\examples{
s <- make_learning_strategy(
  partner_selection = success_bias_select_teacher,
  interaction = success_bias_interact,
  label = "Success-biased"
)
}
