% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{make_model_dynamics}
\alias{make_model_dynamics}
\title{Factory function for creating a ModelDynamics}
\usage{
make_model_dynamics(
  partner_selection,
  interaction,
  model_step = NULL,
  label = "unlabelled"
)
}
\arguments{
\item{partner_selection}{Function to select a partner for any \code{focal_agent} in the \code{model}.}

\item{interaction}{Function for interaction between any \code{focal_agent} and \code{partner} in the \code{model}.}

\item{model_step}{Step function for \code{model} run after all agents select partner and interact.}

\item{label}{Character label for this strategy.}
}
\value{
A \code{ModelDynamics} instance containing social update functions and a metadata-friendly label.
}
\description{
Factory function for creating a ModelDynamics
}
\examples{
success_bias_strategy <- make_model_dynamics(
  partner_selection = success_bias_select_teacher,
  interaction = success_bias_interact,
  label = "Success-biased"
)
# Mock a partner selection, interaction, and model step to show custom use.
mock_selection <- function (focal_agent) NULL
mock_interaction <- function (focal_agent, partner, model) NULL
mock_model_step <- function (model) NULL
mock_strategy <- make_model_dynamics(mock_selection, mock_interaction,
                                        mock_model_step, label = "mock")
# Note make_model_dynamics wraps the R6 class constructor:
mock_strategy_2 <- ModelDynamics$new(mock_selection, mock_interaction,
                                        mock_model_step, mock_strategy) 

}
