% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{make_model_parameters}
\alias{make_model_parameters}
\title{Wrapper for initializing new ModelParameters instance.}
\usage{
make_model_parameters(
  model_dynamics = success_bias_model_dynamics,
  graph = NULL,
  n_agents = NULL,
  ...
)
}
\arguments{
\item{model_dynamics}{Learning strategy to use; must be type LearningStrategy}

\item{graph}{Graph object to use; must inherit igraph}

\item{n_agents}{Number of agents in the model}

\item{...}{Additional model parameters}
}
\description{
Wrapper for initializing new ModelParameters instance.
}
\examples{
# example code

}
