% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{make_preferential_attachment}
\alias{make_preferential_attachment}
\title{Make a preferential attachment network.}
\usage{
make_preferential_attachment(N)
}
\arguments{
\item{N}{population size}
}
\value{
igraph::graph
}
\description{
Make a simple preferential attachment network with N nodes, starting from
two nodes and adding one node per network construction step.
}
