% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{make_small_world}
\alias{make_small_world}
\title{Make a small world network.}
\usage{
make_small_world(N, k, p)
}
\arguments{
\item{N}{Population size}

\item{k}{Seed lattice degree}

\item{p}{Rewire probability}
}
\value{
igraph::graph
}
\description{
Create a small-world network by rewiring a lattice of size N, degree k.
}
