% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{ModelDynamics}
\alias{ModelDynamics}
\title{ModelDynamics}
\description{
ModelDynamics

ModelDynamics
}
\details{
Represents a reusable learning strategy composed of a partner selection function,
an interaction function, and a descriptive label.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelDynamics-new}{\code{ModelDynamics$new()}}
\item \href{#method-ModelDynamics-get_partner_selection}{\code{ModelDynamics$get_partner_selection()}}
\item \href{#method-ModelDynamics-get_interaction}{\code{ModelDynamics$get_interaction()}}
\item \href{#method-ModelDynamics-get_model_step}{\code{ModelDynamics$get_model_step()}}
\item \href{#method-ModelDynamics-get_label}{\code{ModelDynamics$get_label()}}
\item \href{#method-ModelDynamics-clone}{\code{ModelDynamics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelDynamics-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelDynamics-new}{}}}
\subsection{Method \code{new()}}{
Create a new ModelDynamics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDynamics$new(partner_selection, interaction, model_step, label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{partner_selection}}{Function to select the teacher (or NULL)}

\item{\code{interaction}}{Function to define the interaction (or NULL)}

\item{\code{label}}{A string label for this strategy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelDynamics-get_partner_selection"></a>}}
\if{latex}{\out{\hypertarget{method-ModelDynamics-get_partner_selection}{}}}
\subsection{Method \code{get_partner_selection()}}{
Get the partner selection function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDynamics$get_partner_selection()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelDynamics-get_interaction"></a>}}
\if{latex}{\out{\hypertarget{method-ModelDynamics-get_interaction}{}}}
\subsection{Method \code{get_interaction()}}{
Get the interaction function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDynamics$get_interaction()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelDynamics-get_model_step"></a>}}
\if{latex}{\out{\hypertarget{method-ModelDynamics-get_model_step}{}}}
\subsection{Method \code{get_model_step()}}{
Get the interaction function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDynamics$get_model_step()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelDynamics-get_label"></a>}}
\if{latex}{\out{\hypertarget{method-ModelDynamics-get_label}{}}}
\subsection{Method \code{get_label()}}{
Get the strategy label
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDynamics$get_label()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelDynamics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelDynamics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelDynamics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
