% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ModelParameters}
\alias{ModelParameters}
\title{Overwrite existing auxiliary parameters.}
\description{
Overwrite existing auxiliary parameters.

Overwrite existing auxiliary parameters.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelParameters-new}{\code{ModelParameters$new()}}
\item \href{#method-ModelParameters-get_model_dynamics}{\code{ModelParameters$get_model_dynamics()}}
\item \href{#method-ModelParameters-set_model_dynamics}{\code{ModelParameters$set_model_dynamics()}}
\item \href{#method-ModelParameters-get_graph}{\code{ModelParameters$get_graph()}}
\item \href{#method-ModelParameters-set_graph}{\code{ModelParameters$set_graph()}}
\item \href{#method-ModelParameters-get_n_agents}{\code{ModelParameters$get_n_agents()}}
\item \href{#method-ModelParameters-set_n_agents}{\code{ModelParameters$set_n_agents()}}
\item \href{#method-ModelParameters-get_auxiliary}{\code{ModelParameters$get_auxiliary()}}
\item \href{#method-ModelParameters-set_auxiliary}{\code{ModelParameters$set_auxiliary()}}
\item \href{#method-ModelParameters-add_auxiliary}{\code{ModelParameters$add_auxiliary()}}
\item \href{#method-ModelParameters-as_list}{\code{ModelParameters$as_list()}}
\item \href{#method-ModelParameters-clone}{\code{ModelParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelParameters-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelParameters$new(
  model_dynamics = success_bias_strategy,
  graph = NULL,
  n_agents = NULL,
  auxiliary = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelParameters-get_model_dynamics"></a>}}
\if{latex}{\out{\hypertarget{method-ModelParameters-get_model_dynamics}{}}}
\subsection{Method \code{get_model_dynamics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelParameters$get_model_dynamics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelParameters-set_model_dynamics"></a>}}
\if{latex}{\out{\hypertarget{method-ModelParameters-set_model_dynamics}{}}}
\subsection{Method \code{set_model_dynamics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelParameters$set_model_dynamics(model_dynamics)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelParameters-get_graph"></a>}}
\if{latex}{\out{\hypertarget{method-ModelParameters-get_graph}{}}}
\subsection{Method \code{get_graph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelParameters$get_graph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelParameters-set_graph"></a>}}
\if{latex}{\out{\hypertarget{method-ModelParameters-set_graph}{}}}
\subsection{Method \code{set_graph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelParameters$set_graph(graph)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelParameters-get_n_agents"></a>}}
\if{latex}{\out{\hypertarget{method-ModelParameters-get_n_agents}{}}}
\subsection{Method \code{get_n_agents()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelParameters$get_n_agents()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelParameters-set_n_agents"></a>}}
\if{latex}{\out{\hypertarget{method-ModelParameters-set_n_agents}{}}}
\subsection{Method \code{set_n_agents()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelParameters$set_n_agents(n_agents)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelParameters-get_auxiliary"></a>}}
\if{latex}{\out{\hypertarget{method-ModelParameters-get_auxiliary}{}}}
\subsection{Method \code{get_auxiliary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelParameters$get_auxiliary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelParameters-set_auxiliary"></a>}}
\if{latex}{\out{\hypertarget{method-ModelParameters-set_auxiliary}{}}}
\subsection{Method \code{set_auxiliary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelParameters$set_auxiliary(params)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self silently
Add a key-value pair to the auxiliary
variables.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelParameters-add_auxiliary"></a>}}
\if{latex}{\out{\hypertarget{method-ModelParameters-add_auxiliary}{}}}
\subsection{Method \code{add_auxiliary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelParameters$add_auxiliary(key, value)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self silently
Get all parameter values as list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelParameters-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-ModelParameters-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelParameters$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of parameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
