% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.R
\name{new_behavior_observer}
\alias{new_behavior_observer}
\title{Call this to get a new Observer for behaviors.}
\usage{
new_behavior_observer(observer_label = "behavior")
}
\description{
Call this to get a new Observer for behaviors.
}
