% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.R
\name{new_opinion_observer}
\alias{new_opinion_observer}
\title{Call this to get a new Observer for opinions.}
\usage{
new_opinion_observer(observer_label = "opinion")
}
\description{
Call this to get a new Observer for opinions.
}
