% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{not_adjacent}
\alias{not_adjacent}
\title{Check two vertices are not adjacent.}
\usage{
not_adjacent(g, v1, v2)
}
\arguments{
\item{g}{Graph}

\item{v1}{Vertex/agent/node 1}

\item{v2}{Vertex/agent/node 2}
}
\value{
bool indicating whether two vertices v1 and v2 are \emph{not} adjacent in g
}
\description{
Check two vertices are not adjacent.
}
