% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.R
\name{Observer}
\alias{Observer}
\title{Wrap observer function with a label}
\description{
Wrap observer function with a label

Wrap observer function with a label
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Observer-new}{\code{Observer$new()}}
\item \href{#method-Observer-observe}{\code{Observer$observe()}}
\item \href{#method-Observer-clone}{\code{Observer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Observer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Observer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observer$new(observe_fn, label = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Observer-observe"></a>}}
\if{latex}{\out{\hypertarget{method-Observer-observe}{}}}
\subsection{Method \code{observe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observer$observe(model, step, obs_label = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Observer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Observer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
