% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\docType{data}
\name{opinion_dynamics}
\alias{opinion_dynamics}
\title{Basic model dynamics for opinions}
\format{
An object of class \code{ModelDynamics} (inherits from \code{R6}) of length 7.

An object of class \code{ModelDynamics} (inherits from \code{R6}) of length 7.
}
\usage{
opinion_dynamics

opinion_dynamics
}
\description{
Basic model dynamics for opinions

Basic model dynamics for opinions
}
\keyword{datasets}
