% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-w-opinions.R
\name{OpinionAgent}
\alias{OpinionAgent}
\title{Opinion agent extends base socmod::Agent to add opinions and stubbornness}
\description{
Opinion agent extends base socmod::Agent to add opinions and stubbornness

Opinion agent extends base socmod::Agent to add opinions and stubbornness
}
\section{Super class}{
\code{\link[socmod:Agent]{socmod::Agent}} -> \code{OpinionAgent}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OpinionAgent-step_opinions}{\code{OpinionAgent$step_opinions()}}
\item \href{#method-OpinionAgent-new}{\code{OpinionAgent$new()}}
\item \href{#method-OpinionAgent-clone}{\code{OpinionAgent$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="add_neighbors"><a href='../../socmod/html/Agent.html#method-Agent-add_neighbors'><code>socmod::Agent$add_neighbors()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="advance_behavior"><a href='../../socmod/html/Agent.html#method-Agent-advance_behavior'><code>socmod::Agent$advance_behavior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="advance_fitness"><a href='../../socmod/html/Agent.html#method-Agent-advance_fitness'><code>socmod::Agent$advance_fitness()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="degree"><a href='../../socmod/html/Agent.html#method-Agent-degree'><code>socmod::Agent$degree()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="get_attribute"><a href='../../socmod/html/Agent.html#method-Agent-get_attribute'><code>socmod::Agent$get_attribute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="get_attributes"><a href='../../socmod/html/Agent.html#method-Agent-get_attributes'><code>socmod::Agent$get_attributes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="get_behavior"><a href='../../socmod/html/Agent.html#method-Agent-get_behavior'><code>socmod::Agent$get_behavior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="get_fitness"><a href='../../socmod/html/Agent.html#method-Agent-get_fitness'><code>socmod::Agent$get_fitness()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="get_id"><a href='../../socmod/html/Agent.html#method-Agent-get_id'><code>socmod::Agent$get_id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="get_name"><a href='../../socmod/html/Agent.html#method-Agent-get_name'><code>socmod::Agent$get_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="get_neighbors"><a href='../../socmod/html/Agent.html#method-Agent-get_neighbors'><code>socmod::Agent$get_neighbors()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="get_next_behavior"><a href='../../socmod/html/Agent.html#method-Agent-get_next_behavior'><code>socmod::Agent$get_next_behavior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="get_next_fitness"><a href='../../socmod/html/Agent.html#method-Agent-get_next_fitness'><code>socmod::Agent$get_next_fitness()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="remove_neighbors"><a href='../../socmod/html/Agent.html#method-Agent-remove_neighbors'><code>socmod::Agent$remove_neighbors()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="set_attribute"><a href='../../socmod/html/Agent.html#method-Agent-set_attribute'><code>socmod::Agent$set_attribute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="set_attributes"><a href='../../socmod/html/Agent.html#method-Agent-set_attributes'><code>socmod::Agent$set_attributes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="set_behavior"><a href='../../socmod/html/Agent.html#method-Agent-set_behavior'><code>socmod::Agent$set_behavior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="set_fitness"><a href='../../socmod/html/Agent.html#method-Agent-set_fitness'><code>socmod::Agent$set_fitness()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="set_name"><a href='../../socmod/html/Agent.html#method-Agent-set_name'><code>socmod::Agent$set_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="set_neighbors"><a href='../../socmod/html/Agent.html#method-Agent-set_neighbors'><code>socmod::Agent$set_neighbors()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="set_next_behavior"><a href='../../socmod/html/Agent.html#method-Agent-set_next_behavior'><code>socmod::Agent$set_next_behavior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="socmod" data-topic="Agent" data-id="set_next_fitness"><a href='../../socmod/html/Agent.html#method-Agent-set_next_fitness'><code>socmod::Agent$set_next_fitness()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpinionAgent-step_opinions"></a>}}
\if{latex}{\out{\hypertarget{method-OpinionAgent-step_opinions}{}}}
\subsection{Method \code{step_opinions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpinionAgent$step_opinions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpinionAgent-new"></a>}}
\if{latex}{\out{\hypertarget{method-OpinionAgent-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpinionAgent$new(
  id,
  cultural_complexity = 2,
  bounded = FALSE,
  max_op_mag = 1,
  init_op_mean = 0,
  init_op_sd = 1,
  init_opinions = NULL,
  alpha = 1,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpinionAgent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OpinionAgent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpinionAgent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
