% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{plot_homophilynet}
\alias{plot_homophilynet}
\title{Plot homophily network where groups color nodes.}
\usage{
plot_homophilynet(net, node_size = 3, line_width = 0.1, theme_base_size = 10)
}
\arguments{
\item{net}{Homophily network to plot.}

\item{node_size}{Node size in plot.}

\item{line_width}{Edge line width.}

\item{theme_base_size}{Base font size to pass to ggplot theme.}
}
\value{
ggplot
}
\description{
Plot homophily network where groups color nodes.
}
\examples{
hnet_5grp <- make_homophily_network(rep(5, 5), mean_degree = 2, homophily = -0.5)
plot_homophilynet(hnet_5grp)
}
