% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot a summary of behavior adoption over time across trials}
\usage{
plot_summary(summary_df)
}
\arguments{
\item{summary_df}{A tibble from summarise_adoption()}
}
\value{
A ggplot object
}
\description{
Plot a summary of behavior adoption over time across trials
}
