% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run an agent-based model for either a number of iterations (\code{max_its}) or
until some stop condition is met (e.g., fixation).}
\usage{
run(
  model,
  max_its = 1,
  partner_selection = partner_selection_default,
  interaction = interaction_default,
  iterate_model = iterate_model_default,
  stop_cond = stop_cond_default
)
}
\arguments{
\item{model}{}

\item{max_its}{}

\item{partner_selection}{}

\item{interaction}{}

\item{iterate_model}{}

\item{stop_cond}{}
}
\description{
Run an agent-based model for either a number of iterations (\code{max_its}) or
until some stop condition is met (e.g., fixation).
}
