% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{social_influence}
\alias{social_influence}
\title{Bivalent social influence}
\usage{
social_influence(focal_agent, partner, model)

social_influence(focal_agent, partner, model)
}
\arguments{
\item{focal_agent}{}

\item{partner}{}

\item{model}{}
}
\description{
negative = anti-consensus/repulsion
positive = consensus
zero = ignore

negative = anti-consensus/repulsion
positive = consensus
zero = ignore
}
