% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\docType{data}
\name{SOCMOD_PALETTE_CVD}
\alias{SOCMOD_PALETTE_CVD}
\title{CVD-safe custom color palette for scientific plots}
\format{
An object of class \code{character} of length 9.
}
\usage{
SOCMOD_PALETTE_CVD
}
\value{
A named character vector of hex color codes
}
\description{
Recommended for use in \code{scale_color_manual()}.
}
\keyword{datasets}
