% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{step_learning}
\alias{step_learning}
\title{Yet another potential alias for iterate_learning_model to match step_opinions}
\usage{
step_learning(model)
}
\description{
Yet another potential alias for iterate_learning_model to match step_opinions
}
