% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{success_bias_interact}
\alias{success_bias_interact}
\title{Success-biased interaction function}
\usage{
success_bias_interact(learner, teacher, model)
}
\description{
Success-biased interaction function
}
