% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\docType{data}
\name{success_bias_learning_strategy}
\alias{success_bias_learning_strategy}
\title{Define success-biased learning strategy.}
\format{
An object of class \code{LearningStrategy} (inherits from \code{R6}) of length 7.
}
\usage{
success_bias_learning_strategy
}
\description{
Define success-biased learning strategy.
}
\keyword{datasets}
