% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\docType{data}
\name{success_bias_model_dynamics}
\alias{success_bias_model_dynamics}
\title{Define success-biased learning strategy.}
\format{
An object of class \code{ModelDynamics} (inherits from \code{R6}) of length 7.
}
\usage{
success_bias_model_dynamics
}
\description{
Define success-biased learning strategy.
}
\keyword{datasets}
