% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{well_mixed_selection}
\alias{well_mixed_selection}
\title{select randomly either with self-influence or nah}
\usage{
well_mixed_selection(focal, model)

well_mixed_selection(focal, model)
}
\description{
select randomly either with self-influence or nah

select randomly either with self-influence or nah
}
