% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{AgentBasedModel}
\alias{AgentBasedModel}
\title{\code{socmod} implementation of an agent-based model}
\description{
This class encapsulates the model's agents, the network that structures
their interactions, static or dynamic model parameters, model
state, and model outputs.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{agents}}{Named list of agents}

\item{\code{n_agents}}{Number of agents}

\item{\code{step}}{Time step integer index}

\item{\code{network}}{Social network that structures agent interactions}

\item{\code{params}}{Model parameters for use in partner selection,
interaction functions, and}

\item{\code{output}}{Model output, which is either initiated or appended to
depending on if the model has been run yet.
Create a new agent-based model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AgentBasedModel-new}{\code{AgentBasedModel$new()}}
\item \href{#method-AgentBasedModel-add_agents}{\code{AgentBasedModel$add_agents()}}
\item \href{#method-AgentBasedModel-get_agent}{\code{AgentBasedModel$get_agent()}}
\item \href{#method-AgentBasedModel-clone}{\code{AgentBasedModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-new}{}}}
\subsection{Method \code{new()}}{
Agent-based models may be initialized one of three ways: (1)
by providing the agents and the social network that structures their
interaction; (2) providing only the social network; or (3) provide
the number of agents only, which will create and add them to a
complete graph.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$new(agents = NULL, network = NULL, n_agents = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{agents}}{List of agents in agent-based model.}

\item{\code{network}}{igraph Graph instance representing the ABM social network}

\item{\code{n_agents}}{Number of agents.]}

\item{\code{...}}{Additional kwargs that become entries in the AgentBasedModel$params field}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self New agent-based model initialized as specified.
Add agents to the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-add_agents"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-add_agents}{}}}
\subsection{Method \code{add_agents()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$add_agents(agents_to_add)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{agents_to_add}}{Insert a list of new agents into list of existing agents.
Get an agent from the model by name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-get_agent"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-get_agent}{}}}
\subsection{Method \code{get_agent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$get_agent(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of agent, usually an integer or string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
