% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{AgentBasedModel}
\alias{AgentBasedModel}
\title{AgentBasedModel class for socmod}
\description{
AgentBasedModel class for socmod

AgentBasedModel class for socmod
}
\details{
This class represents the main simulation container, holding agents and a social network.
It ensures synchronization between the agent states and the underlying igraph network.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AgentBasedModel-new}{\code{AgentBasedModel$new()}}
\item \href{#method-AgentBasedModel-sync_network}{\code{AgentBasedModel$sync_network()}}
\item \href{#method-AgentBasedModel-get_agent}{\code{AgentBasedModel$get_agent()}}
\item \href{#method-AgentBasedModel-get_network}{\code{AgentBasedModel$get_network()}}
\item \href{#method-AgentBasedModel-get_parameters}{\code{AgentBasedModel$get_parameters()}}
\item \href{#method-AgentBasedModel-set_parameters}{\code{AgentBasedModel$set_parameters()}}
\item \href{#method-AgentBasedModel-set_parameter}{\code{AgentBasedModel$set_parameter()}}
\item \href{#method-AgentBasedModel-get_parameter}{\code{AgentBasedModel$get_parameter()}}
\item \href{#method-AgentBasedModel-clone}{\code{AgentBasedModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize a model with either a graph, agents, or n_agents
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$new(graph = NULL, agents = NULL, n_agents = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{graph}}{An igraph object (optional)}

\item{\code{agents}}{A list of Agent objects (optional)}

\item{\code{n_agents}}{Integer number of agents to create (optional)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-sync_network"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-sync_network}{}}}
\subsection{Method \code{sync_network()}}{
Synchronize agent and network fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$sync_network(
  direction = c("to_graph", "from_graph", "neighbors_only")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{direction}}{"to_graph", "from_graph", or "neighbors_only"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-get_agent"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-get_agent}{}}}
\subsection{Method \code{get_agent()}}{
Get the agent associated with a given ID or name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$get_agent(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Integer index or character name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-get_network"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-get_network}{}}}
\subsection{Method \code{get_network()}}{
Return the igraph network (after syncing from agents)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$get_network()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-get_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-get_parameters}{}}}
\subsection{Method \code{get_parameters()}}{
Get the list of model parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$get_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-set_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-set_parameters}{}}}
\subsection{Method \code{set_parameters()}}{
Set multiple model parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$set_parameters(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{Named list of parameters to set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-set_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-set_parameter}{}}}
\subsection{Method \code{set_parameter()}}{
Set a single model parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$set_parameter(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Parameter name}

\item{\code{value}}{Parameter value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-get_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-get_parameter}{}}}
\subsection{Method \code{get_parameter()}}{
Get a single model parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$get_parameter(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Parameter name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentBasedModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AgentBasedModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentBasedModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
