% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{contagion_model_step}
\alias{contagion_model_step}
\title{Contagion model step}
\usage{
contagion_model_step(model)
}
\arguments{
\item{model}{An AgentBasedModel instance with parameter "drop_rate".}
}
\value{
None. Updates agent behaviors.
}
\description{
Updates all agents for dropping behavior and advances model state.
}
\examples{
model <- abm$new(n_agents = 10, parameters = list(drop_rate = 0.5))
contagion_model_step(model)
}
