% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{contagion_model_step}
\alias{contagion_model_step}
\title{Contagion model step}
\usage{
contagion_model_step(model)
}
\arguments{
\item{model}{An AgentBasedModel instance with parameter "drop_rate".}
}
\value{
None. Updates agent behaviors.
}
\description{
Updates all agents for dropping behavior and advances model state.
}
\examples{
model_parameters <- make_model_parameters(
  learning_strategy = contagion_learning_strategy,
  n_agents = 10, adoption_rate = 0.5, drop_rate = 0.2
)
model <- make_abm(
contagion_model_step(model)
}
