% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{contagion_partner_selection}
\alias{contagion_partner_selection}
\title{Contagion-based partner selection}
\usage{
contagion_partner_selection(learner, model)
}
\arguments{
\item{learner}{An Agent instance.}

\item{model}{An AgentBasedModel instance.}
}
\value{
A single neighbor Agent.
}
\description{
Selects one neighbor at random for potential contagion interaction.
}
\examples{
model <- example_model_with_params()
learner <- model$get_agent(1)
contagion_partner_selection(learner, model)
}
