% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{frequency_bias_interact}
\alias{frequency_bias_interact}
\title{Interaction function for frequency-biased adaptive learning.}
\usage{
frequency_bias_interact(learner, ., model)
}
\arguments{
\item{learner}{Agent currently selected as learner.}

\item{.}{No selected teacher with frequency-biased learning.}

\item{model}{Model variable will be the agent-based model.}
}
\description{
Interaction function for frequency-biased adaptive learning.
}
