% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{make_abm}
\alias{make_abm}
\title{Helper function to create a new agent-based model}
\usage{
make_abm(model_parameters = NULL, agents = NULL)
}
\arguments{
\item{model_parameters}{ModelParameters instance specifying context}

\item{agents}{List of \code{Agent}s}
}
\value{
AgentBasedModel with specified parameters and agents
}
\description{
Helper function to create a new agent-based model
}
\examples{
abm <- make_abm(make_model_parameters(n_agents = 10))
}
