% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parameters.R
\name{make_model_parameters}
\alias{make_model_parameters}
\title{Wrapper for initializing new ModelParameters instance.}
\usage{
make_model_parameters(
  learning_strategy = NULL,
  graph = NULL,
  n_agents = NULL,
  ...
)
}
\description{
Wrapper for initializing new ModelParameters instance.
}
