% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{make_small_world}
\alias{make_small_world}
\title{Make a small world network.}
\usage{
make_small_world(N, k, p, label_func = .sw_label_func)
}
\arguments{
\item{N}{Population size}

\item{k}{Seed lattice degree}

\item{p}{Rewire probability}

\item{label_func}{Specify a function that takes N, k, p and makes a label assigned to igraph via igraph::graph_attr(...)}
}
\value{
igraph::graph
}
\description{
Create a small-world network by rewiring a lattice of size N, degree k.
}
