% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{Neighbors}
\alias{Neighbors}
\title{Neighbors class}
\description{
Neighbors class

Neighbors class
}
\details{
A container for an agent's neighbors, with utilities for iteration, filtering,
sampling, and modification.
}
\examples{

## ------------------------------------------------
## Method `Neighbors$filter`
## ------------------------------------------------

nbrs$filter(\(a) a$get_fitness() > 1)

## ------------------------------------------------
## Method `Neighbors$sample`
## ------------------------------------------------

nbrs$sample(2)
nbrs$sample(weights = \(a) a$get_fitness())

## ------------------------------------------------
## Method `Neighbors$add`
## ------------------------------------------------

nbrs$add(agent1, agent2)

## ------------------------------------------------
## Method `Neighbors$remove`
## ------------------------------------------------

nbrs$remove(agent2)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Neighbors-new}{\code{Neighbors$new()}}
\item \href{#method-Neighbors-map}{\code{Neighbors$map()}}
\item \href{#method-Neighbors-length}{\code{Neighbors$length()}}
\item \href{#method-Neighbors-get}{\code{Neighbors$get()}}
\item \href{#method-Neighbors-filter}{\code{Neighbors$filter()}}
\item \href{#method-Neighbors-sample}{\code{Neighbors$sample()}}
\item \href{#method-Neighbors-add}{\code{Neighbors$add()}}
\item \href{#method-Neighbors-remove}{\code{Neighbors$remove()}}
\item \href{#method-Neighbors-clone}{\code{Neighbors$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neighbors-new"></a>}}
\if{latex}{\out{\hypertarget{method-Neighbors-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neighbors$new(agents = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neighbors-map"></a>}}
\if{latex}{\out{\hypertarget{method-Neighbors-map}{}}}
\subsection{Method \code{map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neighbors$map(f)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neighbors-length"></a>}}
\if{latex}{\out{\hypertarget{method-Neighbors-length}{}}}
\subsection{Method \code{length()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neighbors$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neighbors-get"></a>}}
\if{latex}{\out{\hypertarget{method-Neighbors-get}{}}}
\subsection{Method \code{get()}}{
Get a neighbor by index or name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neighbors$get(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{An integer index or character name of the neighbor}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An Agent instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neighbors-filter"></a>}}
\if{latex}{\out{\hypertarget{method-Neighbors-filter}{}}}
\subsection{Method \code{filter()}}{
Filter neighbors using a predicate function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neighbors$filter(predicate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predicate}}{A function taking an Agent and returning TRUE or FALSE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Neighbors object containing only the filtered agents
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{nbrs$filter(\(a) a$get_fitness() > 1)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neighbors-sample"></a>}}
\if{latex}{\out{\hypertarget{method-Neighbors-sample}{}}}
\subsection{Method \code{sample()}}{
Sample one or more neighbors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neighbors$sample(n = 1, replace = FALSE, weights = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of neighbors to sample}

\item{\code{replace}}{Logical: sample with replacement?}

\item{\code{weights}}{Either a numeric vector or a function returning weights for each agent}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Neighbors object or a single Agent depending on \code{n}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{nbrs$sample(2)
nbrs$sample(weights = \(a) a$get_fitness())
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neighbors-add"></a>}}
\if{latex}{\out{\hypertarget{method-Neighbors-add}{}}}
\subsection{Method \code{add()}}{
Add one or more agents to the neighbors list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neighbors$add(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Agent objects to add}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{nbrs$add(agent1, agent2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neighbors-remove"></a>}}
\if{latex}{\out{\hypertarget{method-Neighbors-remove}{}}}
\subsection{Method \code{remove()}}{
Remove one or more agents from the neighbors list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neighbors$remove(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Agent objects to remove}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{nbrs$remove(agent2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neighbors-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Neighbors-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neighbors$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
