% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plot_homophilynet}
\alias{plot_homophilynet}
\title{Plot homophily network where groups color nodes.}
\usage{
plot_homophilynet(net, node_size = 3)
}
\arguments{
\item{net}{}
}
\description{
Plot homophily network where groups color nodes.
}
