% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_prevalence}
\alias{plot_prevalence}
\title{Plot adoption counts of selected behaviors over time
Plot adoption counts of selected behaviors
(\code{tracked_behaviors}) over time.}
\usage{
plot_prevalence(trials_or_tibble, tracked_behaviors = c("Adaptive"))
}
\arguments{
\item{tracked_behaviors}{Character vector of behaviors to track (e.g., c("Adaptive", "Legacy"))}

\item{trial}{A Trial object}
}
\value{
A ggplot object
}
\description{
Plot adoption counts of selected behaviors over time
Plot adoption counts of selected behaviors
(\code{tracked_behaviors}) over time.
}
