% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{run_trial}
\alias{run_trial}
\title{Run a Trial on an AgentBasedModel with standard learning loop}
\usage{
run_trial(
  model,
  partner_selection = success_bias_select_teacher,
  interaction = success_bias_interact,
  iterate = iterate_learning_model,
  stop = 50
)
}
\arguments{
\item{model}{An AgentBasedModel}

\item{partner_selection}{Function for selecting interaction partner}

\item{interaction}{Function for modifying agents based on partner}

\item{iterate}{Optional function for updating model state}

\item{stop}{Stopping condition: max steps (int) or predicate function (default: 50)}
}
\value{
A Trial object
}
\description{
Run a Trial on an AgentBasedModel with standard learning loop
}
\examples{
agents <- list(
  Agent$new(name = "1", behavior = "Legacy", fitness = 1),
  Agent$new(name = "2", behavior = "Adaptive", fitness = 4),
  Agent$new(name = "3", behavior = "Legacy", fitness = 1),
  Agent$new(name = "4", behavior = "Legacy", fitness = 2)
)
net <- igraph::make_graph(~ 1-2, 1-3, 1-4, 3-2)
model <- AgentBasedModel$new(agents = agents, network = net)
trial <- run_trial(model, stop = 10)
}
