% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{run_trial}
\alias{run_trial}
\title{Run a Trial on an AgentBasedModel with standard learning loop}
\usage{
run_trial(
  model,
  partner_selection = success_bias_select_teacher,
  interaction = success_bias_interact,
  iterate = iterate_learning_model,
  stop = 50,
  label = NULL,
  target_behavior = "Adaptive"
)
}
\arguments{
\item{model}{An AgentBasedModel}

\item{partner_selection}{Function for selecting interaction partner}

\item{interaction}{Function for modifying agents based on partner}

\item{iterate}{Optional function for updating model state}

\item{stop}{Stopping condition: max steps (int) or predicate function}

\item{label}{Optional label to tag the trial (e.g. "success", "frequency")}

\item{target_behavior}{The behavior treated as "adaptation success". Default is "Adaptive".}
}
\value{
A Trial object
}
\description{
Run a Trial on an AgentBasedModel with standard learning loop
}
\examples{
agents <- list(
  Agent$new(name = "1", behavior = "Legacy", fitness = 1),
  Agent$new(name = "2", behavior = "Adaptive", fitness = 4)
)
net <- igraph::make_graph(~ 1-2)
model <- AgentBasedModel$new(agents = agents, graph = net)
trial <- run_trial(model, stop = 10)
}
