% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{run_trials_grid}
\alias{run_trials_grid}
\title{Run a grid of trial ensembles with parameter metadata}
\usage{
run_trials_grid(
  n_trials_per_param,
  model_generator,
  stop = 10,
  learning_strategies = NULL,
  auxlilary_parameters = list(),
  partner_selection = success_biased_teacher_selection,
  interaction = success_biased_interaction,
  model_iterate = iterate_learning_model
)
}
\arguments{
\item{n_trials_per_param}{Number of trials per parameter combination.}

\item{model_generator}{Returns a new AgentBasedModel instance according to model_parameters.}

\item{stop}{Stopping condition (number or function).}

\item{partner_selection}{Function or list of functions to select learning partners.}

\item{interaction}{Function or list of functions to apply during agent interaction.}

\item{model_iterate}{Function to iterate the model.}
}
\value{
A list of Trial objects
}
\description{
Runs trial ensembles across a parameter grid. All scalar and function-valued parameters
used in model construction or trial dynamics are included in metadata for transparency.
}
