% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socmod.R
\docType{package}
\name{socmod-package}
\alias{socmod}
\alias{socmod-package}
\title{socmod: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://css4s.github.io/socmod/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
