% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socmod.R
\docType{package}
\name{socmod-package}
\alias{socmod}
\alias{socmod-package}
\title{socmod: Social behavior models in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides software tools and conceptual framework for creating computational models of social behavior for sustainability.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://SocSci-for-Sustainability.github.io/socmod/}
}

}
\author{
\strong{Maintainer}: Matt Turner \email{maturner01@gmail.com} (\href{https://orcid.org/0000-0001-7421-4552}{ORCID})

}
\keyword{internal}
