% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\docType{data}
\name{SOCMOD_PALETTE}
\alias{SOCMOD_PALETTE}
\title{Custom color palette for scientific plots}
\format{
An object of class \code{character} of length 9.
}
\usage{
SOCMOD_PALETTE
}
\value{
A character vector of hex color codes
}
\description{
This palette was extracted from a provided source image and refined for use in scientific plots,
emphasizing high contrast and perceptual separability. Recommended for use in \code{scale_color_manual()}.
}
\keyword{datasets}
