% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{success_bias_select_teacher}
\alias{success_bias_select_teacher}
\title{Success-biased teacher selection}
\usage{
success_bias_select_teacher(learner, model)
}
\arguments{
\item{learner}{The focal learner agent}

\item{model}{Agent-based model being run with this specification.}
}
\value{
Agent selected from neighbors as teacher.
}
\description{
The probability of selection of
one of the agent's neighbors is proportional to their fitness relative to
other neighbors.
}
\examples{
net <- make_regular_lattice(10, 4)
model <- AgentBasedModel$new(network = net)
for (agent in model$agents) {
  # Agents need non-zero fitness for success bias to work
  agent$curr_fitness <- 1
}
output <- run(model, 10, 
              interact = success_bias_interact, 
              partner_selection = success_bias_select_teacher,
              iterate_model = iterate_learning_model)

}
