% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{summarise_by_metadata}
\alias{summarise_by_metadata}
\title{Summarise trials by metadata fields}
\usage{
summarise_by_metadata(trials, fields)
}
\arguments{
\item{trials}{A list of Trial objects}

\item{fields}{Character vector of metadata fields to group by}
}
\value{
A data frame with group means of success and steps
}
\description{
Summarise trials by metadata fields
}
