% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{Trial}
\alias{Trial}
\title{Trial class for running a single simulation}
\description{
Trial class for running a single simulation

Trial class for running a single simulation
}
\details{
Represents a single run of an AgentBasedModel over time, with customizable
interaction and update logic. Tracks time-series observations and outcome
measures such as adaptation success and steps to fixation.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Trial-new}{\code{Trial$new()}}
\item \href{#method-Trial-run}{\code{Trial$run()}}
\item \href{#method-Trial-get_observations}{\code{Trial$get_observations()}}
\item \href{#method-Trial-get_outcomes}{\code{Trial$get_outcomes()}}
\item \href{#method-Trial-get_label}{\code{Trial$get_label()}}
\item \href{#method-Trial-clone}{\code{Trial$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trial-new"></a>}}
\if{latex}{\out{\hypertarget{method-Trial-new}{}}}
\subsection{Method \code{new()}}{
Initialize a Trial with a model and functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trial$new(
  model,
  partner_selection = NULL,
  interaction,
  iterate = NULL,
  label = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{An AgentBasedModel instance}

\item{\code{partner_selection}}{Optional partner selection function}

\item{\code{interaction}}{Function defining interaction logic}

\item{\code{iterate}}{Optional iteration update function}

\item{\code{label}}{Optional character label for trial group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trial-run"></a>}}
\if{latex}{\out{\hypertarget{method-Trial-run}{}}}
\subsection{Method \code{run()}}{
Run the model and collect results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trial$run(stop = 50, target_behavior = "Adaptive")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stop}}{Either integer for max steps, or predicate function}

\item{\code{target_behavior}}{The behavior treated as "adaptation success"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trial-get_observations"></a>}}
\if{latex}{\out{\hypertarget{method-Trial-get_observations}{}}}
\subsection{Method \code{get_observations()}}{
Return the observation data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trial$get_observations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trial-get_outcomes"></a>}}
\if{latex}{\out{\hypertarget{method-Trial-get_outcomes}{}}}
\subsection{Method \code{get_outcomes()}}{
Return outcome measures
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trial$get_outcomes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trial-get_label"></a>}}
\if{latex}{\out{\hypertarget{method-Trial-get_label}{}}}
\subsection{Method \code{get_label()}}{
Return the label for this trial (if set)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trial$get_label()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trial-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Trial-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trial$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
